requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.12.0
pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
jsonschema==4.19.2
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6